Refactor kube-object-status exports
