+++ src/extensions/registries/kube-object-status-registry.ts
@@ -1,18 +1,5 @@
-import { KubeObject } from "../renderer-api/k8s-api";
-
+import { KubeObject, KubeObjectStatus } from "../renderer-api/k8s-api";
 import { BaseRegistry } from "./base-registry";
-
-export enum KubeObjectStatusLevel {
-  INFO = 1,
-  WARNING = 2,
-  CRITICAL = 3
-}
-
-export type KubeObjectStatus = {
-  level: number;
-  text: string;
-  timestamp?: string;
-}
 
 export interface KubeObjectStatusRegistration {
   kind: string;
